#  #12439 Change _observers from list to set