--- a/src/twisted/logger/_observer.py
+++ b/src/twisted/logger/_observer.py
@@ -29,7 +29,8 @@ class LogPublisher:
     """
 
     def __init__(self, *observers: ILogObserver) -> None:
-        self._observers = set(observers)
+        self._observers_set = set(observers)
+        self._observers = list(observers)
         self.log = Logger(observer=self)
 
     def addObserver(self, observer: ILogObserver) -> None:
@@ -41,7 +42,8 @@ class LogPublisher:
         if not callable(observer):
             raise TypeError(f"Observer is not callable: {observer!r}")
         if observer not in self._observers:
-            self._observers.add(observer)
+            self._observers_set.add(observer)
+            self._observers.append(observer)
 
     def removeObserver(self, observer: ILogObserver) -> None:
         """
@@ -50,6 +52,7 @@ class LogPublisher:
         @param observer: An L{ILogObserver} to remove.
         """
         try:
+            self._observers_set.remove(observer)
             self._observers.remove(observer)
         except KeyError:
             pass
@@ -100,7 +103,7 @@ class LogPublisher:
         @return: A L{Logger} without the given observer.
         """
         errorPublisher = LogPublisher(
-            *(obs for obs in self._observers if obs is not observer)
+            *(obs for obs in self._observers_set if obs is not observer)
         )
         return Logger(observer=errorPublisher)