@allrob23 2025-03-25
Hi @glyph, do you have any insights on this test failure? Could it be a flaky test? It’s passing in my environment:

Test: test_trace
File: src/twisted/logger/test/test_observer.py

![image](https://github.com/user-attachments/assets/c9031d75-cd43-4482-ac52-85d464870e15)