@adiroiban 2025-03-25
@allrob23 The `twisted/twisted` project is not using `pytest` to run the tests.

The tests are executed using `twisted trial` tool.  [Some docs is here](https://docs.twisted.org/en/stable/development/test-standard.html) 

You can trial directly or use `tox`