@glyph 2025-03-25
Running the test in isolation, I also see it pass, but only once: if I do `trial -u twisted.logger.test.test_observer` it quickly fails (on the second observer). see here:
![image](https://github.com/user-attachments/assets/cefb9cf6-1b94-4e1e-8b4e-d901620a4e45)
```
glyph@gale:~/Projects/Twisted  🏰 Twisted  🔀 allrob23/trunk  🐚 🔑   🐳 🫛  ⮐
  ★ trial -u twisted.logger.test.test_observer
 ↩ Tue Mar 25 14:32:01 PDT 2025
Test Pass 1
twisted.logger.test.test_observer
  LogPublisherTests
    test_addObserver ...                                                   [OK]
    test_addObserverNotCallable ...                                        [OK]
    test_fanOut ...                                                        [OK]
    test_interface ...                                                     [OK]
    test_observerRaises ...                                                [OK]
    test_observerRaisesAndLoggerHatesMe ...                                [OK]
    test_observers ...                                                     [OK]
    test_removeObserver ...                                                [OK]
    test_removeObserverNotRegistered ...                                   [OK]
    test_trace ...                                                         [OK]

-------------------------------------------------------------------------------
Ran 10 tests in 0.002s

PASSED (successes=10)
Test Pass 2
twisted.logger.test.test_observer
  LogPublisherTests
    test_addObserver ...                                                   [OK]
    test_addObserverNotCallable ...                                        [OK]
    test_fanOut ...                                                        [OK]
    test_interface ...                                                     [OK]
    test_observerRaises ...                                                [OK]
    test_observerRaisesAndLoggerHatesMe ...                                [OK]
    test_observers ...                                                     [OK]
    test_removeObserver ...                                                [OK]
    test_removeObserverNotRegistered ...                                   [OK]
    test_trace ...                                                       [FAIL]

===============================================================================
[FAIL]
Traceback (most recent call last):
  File "/Users/glyph/Projects/Twisted/src/twisted/logger/test/test_observer.py", line 191, in test_trace
    self.assertEqual(traces[1], ((publisher, o1),))
  File "/Users/glyph/Projects/Twisted/src/twisted/trial/_synctest.py", line 444, in assertEqual
    super().assertEqual(first, second, msg)
  File "/Library/Frameworks/Python.framework/Versions/3.13/lib/python3.13/unittest/case.py", line 907, in assertEqual
    assertion_func(first, second, msg=msg)
  File "/Library/Frameworks/Python.framework/Versions/3.13/lib/python3.13/unittest/case.py", line 1124, in assertTupleEqual
    self.assertSequenceEqual(tuple1, tuple2, msg, seq_type=tuple)
  File "/Library/Frameworks/Python.framework/Versions/3.13/lib/python3.13/unittest/case.py", line 1095, in assertSequenceEqual
    self.fail(msg)
twisted.trial.unittest.FailTest: Tuples differ: ((<tw[104 chars]ls>.o2 at 0x104fe2980>), (<twisted.logger._obs[106 chars]60>)) != ((<tw[104 chars]ls>.o1 at 0x104fe2b60>),)

First differing element 0:
(<twi[63 chars]ction LogPublisherTests.test_trace.<locals>.o2 at 0x104fe2980>)
(<twi[63 chars]ction LogPublisherTests.test_trace.<locals>.o1 at 0x104fe2b60>)

First tuple contains 1 additional elements.
First extra element 1:
(<twisted.logger._observer.LogPublisher object at 0x104f7d9b0>, <function LogPublisherTests.test_trace.<locals>.o1 at 0x104fe2b60>)

  ((<twisted.logger._observer.LogPublisher object at 0x104f7d9b0>,
-   <function LogPublisherTests.test_trace.<locals>.o2 at 0x104fe2980>),
-  (<twisted.logger._observer.LogPublisher object at 0x104f7d9b0>,
-   <function LogPublisherTests.test_trace.<locals>.o1 at 0x104fe2b60>))
+   <function LogPublisherTests.test_trace.<locals>.o1 at 0x104fe2b60>),)
?                                                                      +


twisted.logger.test.test_observer.LogPublisherTests.test_trace
-------------------------------------------------------------------------------
Ran 10 tests in 0.002s

FAILED (failures=1, successes=9)
```

The behavior here is pretty predictable, which is that by moving observers into a `set()`, their previously deterministic order of execution is now non-deterministic. This seems like a potential problem for unit tests in other parts of Twisted as well; I don't think you can make this change exactly as-is without potentially making large numbers of other tests unreliable. Perhaps we need both a list *and* a `set()`? But then of course it's more work to add and remove observers.

All the more reason that we should really start performance work with a benchmark, since maintaining reliability here adds a potentially complex tradeoff.