@allrob23 2025-03-25
I understand, thanks for the explanation, maybe adding a `sorted()` in the test assertion of the tests will solve the problem of reliability?