@glyph 2025-03-26
> I understand, thanks for the explanation, maybe adding a `sorted()` in the test assertion of the tests will solve the problem of reliability?

It's possible that code was relying upon the ordering of log observers previously, and it's not entirely clear to me that they would have been wrong to do so. If we want to make the ordering non-deterministic, we should probably have some tooling to allow developers to test with an intentionally randomized or at least intentionally-different order (perhaps reversed) before unilaterally changing the behavior.