@allrob23 2025-05-02
About the performace check:

The change in `test_log_publisher_call_dispatch` (from 1.3 ms to 1.4 ms, -6.71%) justifies the significant performance gains in other benchmarks. Notably, `test_log_publisher_add_remove` improved dramatically from 137.2 ms to 2.4 ms (57x faster), and `test_lineReceiver[64]` gained 3.25% (8.5 ms to 8.2 ms). On my machine, the `test_log_publisher_call_dispatch` slowdown does not occur, further highlighting the overall positive impact. @adiroiban what do you think?