@adiroiban 2025-05-03
Sorry for the delay.
I don't know what to say here.

This PR has some improvemtns, but at the same time it contains a regression.

I don't have much experience with writing/handling performance tests.

Maybe there is a defect in the way `test_log_publisher_call_dispatch` is written.

I think that log publishing is much more important than add/remove of observers... and it matters that we have a performance issue here.

My suggestion is try to get more eyes for this PR by asking for feedback over the mailing list https://mail.python.org/mailman3/lists/twisted.python.org/